feat: use js-yaml (already installed)
