+++ package.json
@@ -39,7 +39,6 @@
     "test.watch": "yarn test --watch"
   },
   "dependencies": {
-    "@stoplight/yaml": "^2.2.0",
     "js-yaml": "3.12.1",
     "remark-frontmatter": "^1.3.1",
     "remark-parse": "6.0.x",
@@ -52,8 +51,7 @@
     "@stoplight/types": "4.0.0",
     "@types/js-yaml": "3.12.0",
     "@types/unist": "2.0.3",
-    "typescript": "3.3.3",
-    "yaml-ast-parser": "^0.0.43"
+    "typescript": "3.3.3"
   },
   "lint-staged": {
     "*.{ts,tsx}$": [
